requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-hosts==3.0
django-modeladmin-reorder==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
